Fix comment: move tfm specific code into separate tfm files.
