+++ src/libraries/System.Collections.Immutable/ref/System.Collections.Immutable.cs
@@ -127,9 +127,6 @@ namespace System.Collections.Immutable
         public System.ReadOnlyMemory<T> AsMemory() { throw null; }
         public System.ReadOnlySpan<T> AsSpan() { throw null; }
         public System.ReadOnlySpan<T> AsSpan(int start, int length) { throw null; }
-#if NETCOREAPP3_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        public System.ReadOnlySpan<T> AsSpan(System.Range range) { throw null; }
-#endif        
         public System.Collections.Immutable.ImmutableArray<
 #nullable disable
             TOther
