+++ src/libraries/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableArray_1.cs
@@ -794,22 +794,6 @@ namespace System.Collections.Immutable
         /// <returns>The <see cref="ReadOnlySpan{T}"/> representation of the <see cref="ImmutableArray{T}"/></returns>
         public ReadOnlySpan<T> AsSpan(int start, int length) => new ReadOnlySpan<T>(array, start, length);
 
-#if NETCOREAPP3_0_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        /// <summary>
-        /// Creates a <see cref="ReadOnlySpan{T}"/> over the portion of current <see cref="ImmutableArray{T}"/> based on specified <paramref name="range"/>
-        /// </summary>
-        /// <param name="range">Range in current <see cref="ImmutableArray{T}"/>.</param>
-        /// <returns>The <see cref="ReadOnlySpan{T}"/> representation of the <see cref="ImmutableArray{T}"/></returns>
-        public ReadOnlySpan<T> AsSpan(Range range)
-        {
-            var self = this;
-            self.ThrowNullRefIfNotInitialized();
-
-            (int start, int length) = range.GetOffsetAndLength(self.Length);
-            return new ReadOnlySpan<T>(self.array, start, length);
-        }
-#endif
-
         /// <summary>
         /// Copies the elements of current <see cref="ImmutableArray{T}"/> to an <see cref="Span{T}"/>.
         /// </summary>
