+++ src/libraries/System.Collections.Immutable/tests/System.Collections.Immutable.Tests.csproj
@@ -11,6 +11,7 @@
     <Compile Include="EverythingEqual.cs" />
     <Compile Include="ImmutableArrayExtensionsTest.cs" />
     <Compile Include="ImmutableArrayTest.cs" />
+    <Compile Include="ImmutableArrayTest.netstandard21.cs" Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'netstandard2.1'))" />
     <Compile Include="ImmutableArrayTestBase.cs" />
     <Compile Include="ImmutableArray\ImmutableArray.NonGeneric.Tests.cs" />
     <Compile Include="ImmutableDictionaryBuilderTestBase.cs" />
